Refactoring the index.html into an index.php file.
This allows us to use the PURL API integration and make use of the $visitor variable.

Redesigned the layout and functionality of the scene to include proper linkage out to internal resources

incorporated brand and media assets to help cement the feel of the site per references in brand guide.